split core.internals into directory
